Change int64 type to int type

+++ ai-folder.go
@@ -10,15 +10,15 @@ type aiFolderService struct {
 }
 
 type id struct {
-    Id uint64 `json:"id"`
+    Id uint `json:"id"`
 }
 
 // Change a AI folder location
 func (s *aiFolderService) ChangeFolder(
-    folderId uint64, // Folder id to move
-    dstFolderId uint64, // Folder id to move the folder in
+    folderId uint, // Folder id to move
+    dstFolderId uint, // Folder id to move the folder in
 ) (error) {
-    data := "folder_id=" + strconv.FormatUint(folderId, 10) + "&dest_folder_id=" + strconv.FormatUint(dstFolderId, 10)
+    data := "folder_id=" + strconv.FormatUint(uint64(folderId), 10) + "&dest_folder_id=" + strconv.FormatUint(uint64(dstFolderId), 10)
     resp, body, err := s.apiRequest("POST", s.url + "change_folder/", &data)
     if err != nil {
         return err
@@ -33,9 +33,9 @@ func (s *aiFolderService) ChangeFolder(
 
 // Delete a AI folder
 func (s *aiFolderService) Delete(
-    folderId uint64, // Folder id to delete
+    folderId uint, // Folder id to delete
 ) (error) {
-    data := "folder_id=" + strconv.FormatUint(folderId, 10)
+    data := "folder_id=" + strconv.FormatUint(uint64(folderId), 10)
     resp, body, err := s.apiRequest("POST", s.url + "delete/", &data)
     if err != nil {
         return err
@@ -50,9 +50,9 @@ func (s *aiFolderService) Delete(
 
 // Create a AI folder
 func (s *aiFolderService) New(
-    folderId uint64, // Parent folder id
-) (uint64, error) {
-    data := "folder_id=" + strconv.FormatUint(folderId, 10)
+    folderId uint, // Parent folder id
+) (uint, error) {
+    data := "folder_id=" + strconv.FormatUint(uint64(folderId), 10)
     resp, body, err := s.apiRequest("POST", s.url + "new/", &data)
     if err != nil {
         return 0, err
@@ -73,10 +73,10 @@ func (s *aiFolderService) New(
 
 // Rename a AI folder
 func (s *aiFolderService) Rename(
-    folderId uint64, // Folder id to rename
+    folderId uint, // Folder id to rename
     name string, // The new name of the folder
 ) (error) {
-    data := "folder_id=" + strconv.FormatUint(folderId, 10) + "&new_name=" + name
+    data := "folder_id=" + strconv.FormatUint(uint64(folderId), 10) + "&new_name=" + name
     resp, body, err := s.apiRequest("POST", s.url + "rename/", &data)
     if err != nil {
         return err
